7 - Manage Prices in Django Admin with Tabular Inlines

+++ src/subscriptions/models.py
@@ -75,7 +75,7 @@ class SubscriptionPrice(models.Model):
         """
         remove decimal places
         """
-        return self.price * 100
+        return int(self.price * 100)
 
     @property
     def product_stripe_id(self):
